chargeback tests to validate resource allocation costs

+++ cfme/tests/intelligence/chargeback/test_resource_allocation.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-""" Tests to validate chargeback costs for resource allocation to VMs"""
+""" Tests to validate chargeback costs for allocated resources(memory, cpu, storage) to VMs"""
 
 from datetime import timedelta
 
@@ -53,7 +53,7 @@ def vm_ownership(enable_candu, provider, appliance):
     user = None
     user = appliance.collections.users.create(
         name=provider.name + fauxfactory.gen_alphanumeric(),
-        credential=Credential(principal='uid' + '{}'.format(fauxfactory.gen_alphanumeric()),
+        credential=Credential(principal='uid{}'.format(fauxfactory.gen_alphanumeric()),
             secret='secret'),
         email='abc@example.com',
         groups=cb_group,
@@ -116,9 +116,7 @@ def verify_vm_uptime(appliance, provider):
     """
     vm_name = provider.data['cap_and_util']['chargeback_vm']
     vm_creation_time = appliance.rest_api.collections.vms.get(name=vm_name).created_on
-    if appliance.utc_time() - vm_creation_time > timedelta(hours=1):
-        return True
-    return False
+    return appliance.utc_time() - vm_creation_time > timedelta(hours=1)
 
 
 @pytest.fixture(scope="module")
@@ -126,15 +124,15 @@ def resource_alloc(vm_ownership, appliance, provider):
     """Retrieve resource allocation values"""
     vm_name = provider.data['cap_and_util']['chargeback_vm']
 
-    if provider.type == 'scvmm':
+    if provider.one_of(SCVMMProvider):
         wait_for(verify_vm_uptime, [appliance, provider], timeout=3650,
-            fail_condition=False, message='Waiting for VM to be up for at least one hour')
+            message='Waiting for VM to be up for at least one hour')
 
     vm = appliance.rest_api.collections.vms.get(name=vm_name)
     vm.reload(attributes=['allocated_disk_storage', 'cpu_total_cores', 'ram_size'])
     return {"storage_alloc": vm.allocated_disk_storage,
             "memory_alloc": vm.ram_size,
-            "cpu_alloc": vm.cpu_total_cores}
+            "vcpu_alloc": vm.cpu_total_cores}
 
 
 def resource_cost(appliance, provider, metric_description, usage, description, rate_type):
@@ -143,9 +141,6 @@ def resource_cost(appliance, provider, metric_description, usage, description, r
     details = appliance.db.client['chargeback_rate_details']
     cb_rates = appliance.db.client['chargeback_rates']
     list_of_rates = []
-
-    def add_rate(tiered_rate):
-        list_of_rates.append(tiered_rate)
 
     with appliance.db.client.transaction:
         result = (
@@ -159,7 +154,7 @@ def resource_cost(appliance, provider, metric_description, usage, description, r
     for row in result:
         tiered_rate = {var: getattr(row, var) for var in ['variable_rate', 'fixed_rate', 'start',
             'finish']}
-        add_rate(tiered_rate)
+        list_of_rates.append(tiered_rate)
 
     # Check what tier the usage belongs to and then compute the usage cost based on Fixed and
     # Variable Chargeback rates.
@@ -196,10 +191,10 @@ def chargeback_report_custom(appliance, vm_ownership, assign_custom_rate, provid
     """Create a Chargeback report based on a custom rate; Queue the report"""
     owner = vm_ownership
     data = {
-        'menu_name': 'cb_custom_' + provider.name,
-        'title': 'cb_custom' + provider.name,
+        'menu_name': 'cb_custom_{}'.format(provider.name),
+        'title': 'cb_custom_{}'.format(provider.name),
         'base_report_on': 'Chargeback for Vms',
-        'report_fields': ['Memory Allocated', 'Memory Allocated over Time Period', 'Owner',
+        'report_fields': ['Memory Allocated Cost', 'Memory Allocated over Time Period', 'Owner',
         'vCPUs Allocated over Time Period', 'vCPUs Allocated Cost',
         'Storage Allocated', 'Storage Allocated Cost'],
         'filter': {
@@ -213,14 +208,19 @@ def chargeback_report_custom(appliance, vm_ownership, assign_custom_rate, provid
     logger.info('Queuing chargeback report with custom rate for {} provider'.format(provider.name))
     report.queue(wait_for_finish=True)
 
-    yield list(report.saved_reports.all()[0].data.rows)
-    report.delete()
+    if not list(report.saved_reports.all()[0].data.rows):
+        pytest.skip('Empty report')
+    else:
+        yield list(report.saved_reports.all()[0].data.rows)
+
+    if report.exists:
+        report.delete()
 
 
 @pytest.yield_fixture(scope="module")
 def new_chargeback_rate():
     """Create a new chargeback rate"""
-    desc = 'custom_' + fauxfactory.gen_alphanumeric()
+    desc = 'custom_{}'.format(fauxfactory.gen_alphanumeric())
     compute = rates.ComputeRate(description=desc,
                 fields={'Allocated CPU Count':
                         {'per_time': 'Hourly', 'variable_rate': '2'},
@@ -234,83 +234,61 @@ def new_chargeback_rate():
     storage.create()
     yield desc
 
-    compute.delete()
-    if not BZ(1532368, forced_streams=['5.9']).blocks:
+    if compute.exists:
+        compute.delete()
+    if not BZ(1532368, forced_streams=['5.9']).blocks and storage.exists:
         storage.delete()
 
 
-def generic_test_chargeback_cost(chargeback_costs_custom, chargeback_report_custom, column,
-        resource_alloc_cost, soft_assert):
+@pytest.mark.parametrize('resource_cost',
+                         ['memory', 'vcpu', 'storage'],
+                         ids=['memory_alloc', 'vcpu_alloc', 'storage_alloc'])
+def test_validate_cost(chargeback_costs_custom, chargeback_report_custom, soft_assert,
+        resource_cost):
+
     for groups in chargeback_report_custom:
-        if not groups[column]:
+        if not (groups['Memory Allocated Cost'] or groups['Storage Allocated Cost'] or
+        groups['vCPUs Allocated Cost']):
             pytest.skip('missing column in report')
         else:
-            estimated_resource_alloc_cost = chargeback_costs_custom[resource_alloc_cost]
-            cost_from_report = groups[column]
+            if groups['Memory Allocated Cost']:
+                estimated_resource_alloc_cost = chargeback_costs_custom['memory_alloc_cost']
+            cost_from_report = groups['Memory Allocated Cost']
+
+            if groups['vCPUs Allocated Cost']:
+                estimated_resource_alloc_cost = chargeback_costs_custom['vcpu_alloc_cost']
+            cost_from_report = groups['vCPUs Allocated over Time Period']
+
+            if groups['Storage Allocated Cost']:
+                estimated_resource_alloc_cost = chargeback_costs_custom['storage_alloc_cost']
+            cost_from_report = groups['Storage Allocated Cost']
+
             cost = cost_from_report.replace('$', '').replace(',', '')
             soft_assert(estimated_resource_alloc_cost - DEVIATION <=
                 float(cost) <= estimated_resource_alloc_cost + DEVIATION,
                 'Estimated cost and report cost do not match')
 
 
-def generic_test_resource_alloc(resource_alloc, column, resource, soft_assert):
+@pytest.mark.parametrize('resource',
+                         ['memory', 'vcpu', 'storage'],
+                         ids=['memory_alloc', 'vcpu_alloc', 'storage_alloc'])
+def test_verify_allocation(resource_alloc, resource, soft_assert):
     for groups in chargeback_report_custom:
-        if not groups[column]:
+        if not (groups['Memory Allocated over Time Period'] or groups['Storage Allocated'] or
+        groups['vCPUs Allocated over Time Period']):
             pytest.skip('missing column in report')
         else:
             allocated_resource = resource_alloc[resource]
-            resource_from_report = groups[column].replace('MB', '').replace('GB', '')
-            soft_assert(allocated_resource - DEVIATION <=
-                float(resource_from_report) <= allocated_resource + DEVIATION,
+
+            if groups['Memory Allocated over Time Period']:
+                resource_from_report = groups['Memory Allocated over Time Period'].\
+                    replace('MB', '').replace('GB', '')
+
+            if groups['vCPUs Allocated over Time Period']:
+                resource_from_report = groups['vCPUs Allocated over Time Period']
+
+            if groups['Storage Allocated']:
+                resource_from_report = groups['Storage Allocated']
+
+            soft_assert(allocated_resource == resource_from_report,
                 'Estimated cost and report cost do not match')
-
-
-def test_verify_alloc_memory(resource_alloc, column, resource, soft_assert):
-    """Test to validate cost for memory allocation
-       Calculation is based on custom Chargeback rate.
-    """
-    generic_test_resource_alloc(resource_alloc,
-        'Memory Allocated', 'memory_alloc', soft_assert)
-
-
-def test_verify_alloc_cpu(resource_alloc, column, resource, soft_assert):
-    """Test to validate cost for memory allocation
-       Calculation is based on custom Chargeback rate.
-    """
-    generic_test_resource_alloc(resource_alloc,
-        'vCPUs Allocated over Time Period', 'cpu_alloc', soft_assert)
-
-
-def test_verify_alloc_storage(resource_alloc, column, resource, soft_assert):
-    """Test to validate cost for memory allocation
-       Calculation is based on custom Chargeback rate.
-    """
-    generic_test_resource_alloc(resource_alloc,
-        'Storage Allocated', 'storage_alloc', soft_assert)
-
-
-def test_validate_alloc_memory_cost(chargeback_costs_custom, chargeback_report_custom,
-        soft_assert):
-    """Test to validate cost for memory allocation
-       Calculation is based on custom Chargeback rate.
-    """
-    generic_test_chargeback_cost(chargeback_costs_custom, chargeback_report_custom,
-        'Memory Allocated Cost', 'memory_alloc_cost', soft_assert)
-
-
-def test_validate_alloc_vcpu_cost(chargeback_costs_custom, chargeback_report_custom,
-        soft_assert):
-    """Test to validate cost for vCPU allocation
-       Calculation is based on custom Chargeback rate.
-    """
-    generic_test_chargeback_cost(chargeback_costs_custom, chargeback_report_custom,
-        'vCPUs Allocated Cost', 'vcpu_alloc_cost', soft_assert)
-
-
-def test_validate_alloc_storage_cost(chargeback_costs_custom, chargeback_report_custom,
-        soft_assert):
-    """Test to validate cost for vCPU allocation
-       Calculation is based on custom Chargeback rate.
-    """
-    generic_test_chargeback_cost(chargeback_costs_custom, chargeback_report_custom,
-        'Storage Allocated Cost', 'storage_alloc_cost', soft_assert)
